Refactor oscn DailyFilings and Search to use a BaseSearch class #103
