+++ court_scraper/platforms/oscn/pages/search.py
@@ -1,11 +1,9 @@
-import datetime
 import logging
 import requests
 
-from court_scraper.case_info import CaseInfo
 from court_scraper.utils import dates_for_range
-from .case_number_lookup import CaseNumberLookup
 
+from .base_search import BaseSearch
 from .search_results import SearchResultsPage
 from ..search_results_wrapper import SearchResultsWrapper
 
@@ -13,21 +11,17 @@ from ..search_results_wrapper import SearchResultsWrapper
 logger = logging.getLogger(__name__)
 
 
-class Search:
+class Search(BaseSearch):
     """General search page for all OK counties.
 
     Supports searches by date, case type and a variety of other
     parameters. Large searches are truncated, so searches using
     this class should be targeted narrowly (e.g. a single day
-    for smaller counties).
+    for smaller counties). For larger counties such as Tulsa, 
+    use DailyFilings search class.
 
     Args:
-        - place_id (str): Standard place id (e.g. ok_alfalfa or ok_roger_mills)
-
-    Returns:
-
-        SearchResultsWrapper class containing html of search results page and
-        CaseInfo instances.
+        - place_id (str): Standard place id (e.g. ok_alfalfa)
 
     """
 
@@ -79,28 +73,6 @@ class Search:
         html = response.text
         page = SearchResultsPage(self.place_id, response.text)
         return html, page.results
-
-    def _scrape_case_details(self, date_key, search_results, basic_case_data):
-        """Loop through CaseInfo classes from upstream search results,
-        scrape case details, and merge data for CaseInfo from both sources
-        """
-        for basic_case in basic_case_data:
-            lookup = CaseNumberLookup(self.place_id)
-            # Get single CaseInfo class from CaseNumberLookup
-            case_info = lookup.search([basic_case.number])[0]
-            # Merge basic CaseInfo (from search results page)
-            # with CaseInfo (from case detail page)
-            basic_case.merge(case_info)
-            # ...and add to search results, which expects a list
-            search_results.add_case_data(date_key, [basic_case])
-
-    def _standardize_date(self, date_str, from_format, to_format):
-        return datetime.datetime.strptime(date_str, from_format).strftime(to_format)
-
-    @property
-    def _place(self):
-        county_bits = self.place_id.replace('_', ' ').split(' ')[1:]
-        return " ".join(county_bits).title()
 
     @property
     def _default_params(self):
